refactor: Simplify Autofix.render_fix interface (#6407)
All of the paremeters to `Autofix.render_fix` can be derived from a
`Pattern_match.t`, and it's hard to foresee a use case where a caller
doesn't have a `Pattern_match.t` but still wants a rendered autofix, and
this is a lot simpler.

`Autofix.render_fix` now returns a `Textedit.t` instead of the
replacement text alone. For now, the range in it is just the range of
the `Pattern_match.t`, but that might not always be the case. Maybe
later it will become a `Textedit.t list`.

The only downside is that this change introduces an additional file read
in tests, but that's not a big deal.

Test plan: Automated tests