Make it possible to turn off skipping minimized files (#6415)
* Make it possible to turn off skipping minimized files

See https://github.com/returntocorp/semgrep-proprietary/pull/328

This helps DeepSemgrep. I could instead have added an option to tune the
amount of whitespace, but for now I wanted to err on the side of skipping
fewer files, since performance is not a priority for DeepSemgrep. We can
revisit this later.

Test plan: see the DeepSemgrep PR; nothing should change for Semgrep users

* precommit