#5698 fix(yaml): excessive capture in maps (#6464)
* add test

* fix(yaml): excessive text capture in maps

Yaml-mode captures too much text in mapping contexts; it captures all
text from the final token in the map to the next token in the file. This
includes comments and leading whitespace. This comes from the upstream
library (https://github.com/avsm/ocaml-yaml).

The implemented fix is to use the position of the token in the event
immediately prior to the `Mapping_end` event. If there is no prior
event, fall back to the `Mapping_end` event's position.

Fixes #5698

* slight optimisation, documentation

* add changelog

* pre-commit fixes