Ensure that Rule.formula is serializable
The `Regexp_engine.t` type is a wrapper for a C type coming from the
C PCRE library, and it is represented by a "Custom" block. Values of
this type are not serializable, and will cause an exception if we try
calling `Marshal.to_string` (or variant) on them. This can cause
problems when serializing any data structure that depends on `Xpattern.t`
and `Rule.formula`.

Previously, we got weird crashes after PR #5725 (taint labels). This was
due to redefining type `Taint.source` as `Rule.taint_source call_trace`,
where `Rule.taint_source` contains a `Rule.formula`. Whenever a taint
finding is produced, we compute the taint trace this way:

    lazy (taint_trace_of_src_to_sink source tokens sink)

This `lazy ...` expression creates a thunk that capturies a `source`
which contains a `Rule.formula`. Even though the formula is not used to
compute the taint trace, it still needs to be serialized in order to
serialize the thunk.

So, when using -j 2 (or higher) and -json (in text mode we print findings
right away and hence the thunk gets evaluated), this thunk remains
unevaluated and reaches the result of `Match_rules.check`. Then Parmap
tries to serialize it to communicate the result back to the master
process... and if the formula contains any compiled regex, it crashes.

This also affects us if we try to serialize a `Dataflow_tainting.config`
for caching purposes. While we could remove the formulas from these
types, it is less error prone to just make sure that formulas can be
serialized.

Follows: 2e6b3079d51 ("Quick-fix for regression in 0.107.0 after taint labels PR (#5867)")

Regarding perf, this should not cause any noticeable regression, unless
we build some synthetic case with lots of rules making an absurd use of
regexes. I ran `semgrep -c p/default` on three repos from stress-test-monorepo
(elasticsearch, vscode, and mediawiki) and with this change all runs were
actually a little bit faster (maybe just coincidence).

test plan:
make test
(In PR #5867 we modified Test_engine.ml to check that the result of
 Match_rules.check can be serialized, see 'Test_engine.make_tests'.)