refactor: Add `Require` node to generic AST
Currently, in JS, `require('foo')` is represented using `OtherExpr` for
the `require` part of that expression. This is fine currently since we
don't do any analysis of `require` calls after the AST generic
translation. Instead, we desugar them to `ImportFrom` in some cases, and
Semgrep's existing naming logic for imported symbols takes it from
there.

However, that desugaring is causing some issues. See #6532 for details.

Instead of desugaring, we need to represent `Require` in the AST
properly, and analyze it directly. We could simply treat it like a
normal function call to a function named `require`, but it's a separate
node in the JS AST and doing analysis by matching on specific literal
values is more error-prone.

Test plan: Automated tests