tainting: Assume sources and sanitizers are pure
Pure in the sense of having no side-effects.

In 0.94.0 we made it so when a pattern-source matched a variable exactly
this was understood as that variable being tainted by side-effect, e.g.
given `tainted(x)` we would taint `x` by side-effect. Same was done for
sanitizers.

This worked well initially, or so we thought, until in 0.113.0 we added
field-sensitivity, and in subsequent versions we applied sources and
sanitizers by side-effect on more general l-values. It was then that we
started to see regressions that were fairly unintuitive for users (see
added tests).

We now correct the situation by adding the `by-side-effect` option to
sources and sanitizers, and requiring this option to be explicitly enabled
in order to apply the source or the sanitizer by side-effect.

Requires returntocorp/semgrep-rules#2577

Fixes: 653db7dfa78 ("tainting: Allow variables to be taint sources by side-effect (#5240)")
Fixes: 4ced4711358 ("tainting: Allow variables to be sanitized by side-effect (#5266)")

Closes PA-1629
Closes PA-2229

Reviewed-by: Emma Jin

test plan:
make test # added three tests