Fix(js): Remove desugaring for JS requires (#6636)
## Background

My end goal here is to add CommonJS import/export support to
DeepSemgrep. To do this, I need some consistency in how requires are
represented.

Currently, we desugar `require` calls to `ImportFrom` nodes in one
specific case. This only occurs when the `require` is on the RHS of a
variable assignment, and the LHS is a destructuring, e.g. `const { f1,
f2 } = require("file");`. However, instead of replacing the `require`,
the `ImportFrom` appears alongside it. I believe that this was done
because there isn't a way to consistently transform require calls into
`ImportFrom`. `require` calls are simply expressions, and can appear
anywhere. There is also the potential for complicated destructuring
where Semgrep bails out from desugaring at all. Thus, to be able to
consistently match a pattern such as `require('file')` we need to keep
the original node around.

There is no equivalent desugaring to `ImportAs` for `const f =
require("file");`, thus the TODO item in the `aliasing_require` test
that is addressed by this change.

Because there is no way to map all `require` calls to the existing
import nodes that we have, we need to remove this partial desugaring if
we want to build a maintainable analysis.

## Implementation

This removes the code that does the desugaring in `js_to_generic.ml`.
With that desugaring gone, we need to make sure that both naming and
matching behave as desired, so I've added code to handle `require` calls
to both `Naming_AST.ml` and `Generic_vs_generic.ml`.

Some things for reviewers to focus on:
- I added `m_directive_vs_def`. This performs matching on two different
  kinds of AST nodes. I couldn't find a precedent for this. Is this the
  right thing to do? Thoughts on naming? For matching partial nodes we
  generate synthetic nodes in the visitor. Maybe we should use that
  approach here? I think this is clearer, though.
- Related to the above, I added the `general_matcher` type which takes
  two type parameters so that `m_list_in_any_order` can operate on a
  matcher that matches a different type on the left and the right.

## Test plan

Automated tests