tainting: Fix regression in DeepSemgrep after PR #6640 (#6669)
Better check unconditionally whether the function is a sink and, if so,
check whether its arguments are tainted. When the function acts like a
taint propagator, that will produce some duplicate findings, but they
should be filtered out during deduplication anyways.

In addition, the taint of the function expression itself should always
be propagated, also in DeepSemgrep. The function/method itself could be
a taint source. Plus, for now it helps us identify `x.foo()` as tainted
when `x` is tainted, although in the future we want to do this properly.

Fixes: 1eb330d6887 ("tainting: Make identification of sinks more precise (#6640)")

test plan:
make test