fix: Avoid stack overflow when symbol resolves to itself
This addresses an issue encountered in DeepSemgrep. I don't know how to
trigger this with regular Semgrep.

Test plan: Automated tests, plus new automated tests to be added in
DeepSemgrep.

Also ran it manually with `--debug` in the crashing case I have for
DeepSemgrep and observed the new log lines.