fix: Bind module name metavariables as text (#6702)
Fixes #6674

Now this just falls through to the next case, where the metavariable is
bound as a `Text` kind, rather than an `E` (expression).

I'm not sure why this would make pretty printing more difficult.
However, given that the old pretty printer is used only in experiments,
and the new one doesn't handle this construct yet, I think we should
prioritize fixing the bug.

Test plan:

Automated tests

Manual test on the example in #6674 (tests for this kind of bug do not
fit in well with our existing automated test framework). The error
message is "package match is foo" (note the lack of quotes around
"foo").