fix: Invalidation for file cache in Range.ml
This addresses a DeepSemgrep crash when run on the `NodeBB` repo (see
test plan), but it's probably possible to trigger this crash in other
ways including without DeepSemgrep.

Basically, the issue is that collisions in tmp file names can lead to
outdated information available in the cache. If we try to read, a
substring starting at, say character 10, but the temp file that is
cached is only 5 characters long, we get a crash.

Test plan: See https://github.com/returntocorp/pfff/pull/575