Run p/deep-semgrep rules on all CI runs with the deep flag
This is a temporary hack that should be removed _soon_. What this will
do is, for users with the deep toggle, the p/deepsemgrep rules will be
appended to the rules configured on their rule board.

This will help people trialing deep-semgrep, since all they will have
to do is turn on the toggle to get the correct configuration.

After we figure out the ideal flow for app users, we should delete this
code.

Test plan: run `semgrep ci` on `examples-annotations/repos/java/SQLi/historical-commit-repos/simple-mvc-app`. It should find something, where it wouldn't
have before.