const-prop: Fix var_stats for l-values such as this.x (#6825)
* const-prop: Fix var_stats for l-values such as this.x

In 0.123.0 we allowed const-prop to regard a private class attribute in
Java as constant, if it was assigned just once. But when we looked for
assigned variables we didn't consider l-values such as `this.x`, so
`this.x = 42` was not being counted as an assignment to `x`.

And, refactor code so that it is cleaner plus it will be able to handle
e.g. nested tuples in an LHS.

Closes #6793
Follows: 621e5ea3b1e ("propagate isngly assigned private java vars (#6598)")

test plan:
make test # added one test

* Yoann's review