fix: In constant propagation, use svalues when they already exist (#6839)
As noted inline, in some cases the svalue is computed in DeepSemgrep
ahead of time. When that's the case, we should use the precomputed
svalue instead of looking on the RHS of the assignment.

In particular, this makes my work in progress to better support CJS
require calls work properly. `const x = require('y')` gets resolved
correctly and the svalue for `x` gets set in DeepSemgrep's naming, but
subsequent uses of `x` hit the codepath modified here and end up having
their svalues set to `Sym(require('y')`, which the Semgrep code cannot
resolve.

Test plan: Automated tests (no change)