fix: Make `$F(x)` match `eval(x)` (#6878)
`eval` is an `IdSpecial` in the generic AST, instead of a regular
identifier. However, users expect it to match like an ordinary function
call, since it actually is.

I previously added special cases for `this`, `super`, `self`, `parent`,
and `require`, which are all `IdSpecial`s to make sure that they all
match. I left a TODO for `eval` since it seemed like it should probably
match but was unrelated to what I was doing at the time. This fix simply
addresses that TODO and adds a test.

Fixes #6877

Test plan: Automated tests