fix: Avoid fatal merge-base calls on shallow checkouts
#6756 allows us to use the GitHub API to find the merge base, rather
than relying on the sometimes slow process of fetching additional
commits until we can compute it locally.

However, downstream code assumes that we have enough history to compute
the merge base. When it tries to do so, Git fatals and Semgrep aborts.
This was reported to us when a customer tried to make use of this
optimization.

I missed this while testing the original PR because my test PR on my
test repo added the file with the finding. When all the files with
findings are added in the PR being scanned, Semgrep doesn't bother doing
a baseline scan, which I overlooked.

Test plan:

Same as #6756 except the test PR modifies an existing file with findings
rather than creating a new file. When using the current
`returntocorp/semgrep` image, Semgrep fails on the merge base call in
the same way as described by the customer. When using the locally-built
`nmote/semgrep` image with the change in this PR, Semgrep runs
correctly.