ResolvedName of dotted_ident -> GlobalName of string list
This is the PR for a possible resolution of the
ResolvedName and metavariable mess.
The idea is to not store tokens anymore for resolved names
(now rename to GlobalName), because anyway the tokens
in it can come from different files, or contain tokens whose
range might contain keywords (e.g., in import {bar} from foo).
We want to move away from tokens anyway as a general principle,
and prefer range and proper pretty printing.

test plan:
make test, and hope for no regression in semgrep.
At some point also test in deep-semgrep and test
with the new testsuite brandon is preparing covering
more edge cases of interactions between resolved names
and metavariable-regex: and metavariable-pattern: