metavariable-pattern: Force Spacegrep to analyze requested text
Even if the text matched by a metavariable looks like machine generated,
Spacegrep must analyze it because it has been requested explicitly. This
is text coming from a presumably legit file, so it's not the same as
ignoring an entire file that looks like machine generated.

Closes PA-2386

test plan:
make test # added one test