feat: move docker build later in the release process
In this way, we avoid releasing docker (a less risky release artifact) before other release artifacts (such as macos/ubuntu builds) in case of issues.

This avoids getting into the state where docker image vX.Y.Z is released, but we can't release on other platforms and need to scrap the release.