feat(metrics): add the engine type to the metrics sent
This adds the engine that the user requests to the metrics sent.
Otherwise, for CI scans this information is not present anywhere.

This also moves information that can be sent earlier to be sent
earlier so that even in case of failure we have it.

Test plan: see added tests