feat(js): Allow metavar ellipsis in template strings
Previously, a metavar ellipsis in a template string would match only its
literal value. This is surprising, since ellipses work in template
strings, and metavar ellipses work in regular string literals.

To address this, I updated `m_arguments_concat` to use the
`m_list_with_dots_and_metavar_ellipsis` helper function, instead of
using a hand-rolled solution to match ellipses.

The change in `Pattern.ml` is so that the optimization step in
`Analyze_rule` does not erroneously skip files that don't contain the
literal contents of the metavariable ellipsis. Without this change, a
manual test of the case added in `tests/rules` fails. However, the
automated test runner does not appear to run the optimization step in
the same way, and it still passes. I will file an issue to follow up on
that.

Test plan: Automated tests