feat: Allow metavariable-pattern to be more widely used with generic mode
As I've explained inline in more detail, we now allow
metavariable-pattern to be used on any kind of metavariable binding as
long as the target language for the metavariable-pattern is generic.
Previously we only allowed metavariable-pattern to be used with a sub
language if the metavariable in question was bound to a string literal.

Test plan: Automated tests