fix: Correct range for mvars bound in `metavariable-pattern`
Followup to #7052. It hasn't been released yet, so I'm not adding
anything to the changelog here.

Specifically in the case where the `metavariable-pattern` uses a
different language, the metavariable binding location was not correctly
calculated.

I just pulled out the `revert_loc` and `fix_loc` functions into an outer
scope, so that `revert_loc` is also available in the case where the
`metavariable-pattern` is evaluated using a different language.

I added some TODOs inline because I'm suspicious of the way that the
location are adjusted. However, since `fix_loc` predates #7502 and
`revert_loc` is simply the inverse, I don't want to mess with it right
now.

Test plan: Automated tests. Without this change, the added test has no
matches, since the `focus-metavariable` range doesn't intersect with the
original match's range.