feat(ci): allow users to run an individual repo in CI with a different engine (#7095)
* feat(ci): allow users to run an individual repo in CI with a different engine

The CLI allows users to choose between running with just pro languages, with
intrafile interprocedural analysis, and with interfile analysis. We offer this
to CI users too in case they wish to make an exception for a single repo. In
general, they should use the toggle to control this.

Test plan: run in CI with docker image

* Commands should override the toggle

* Update the message