fix: Fix location reported by nested metavariables
When we evaluate a `metavariable-pattern` we make a temp file with the
original contents. We adjust each location in the original AST to match
the temp file. When we get metavariable bindings out, we need to adjust
their locations back to the original file.

We only need to adjust the column for locations that are on the first
line of the original metavariable match. For subsequent lines, the
column will already be correct. Previously we were blindly adjusting the
column, leading to the incorrect results outlined in the test plan
below.

This does not appear to affect matching in any way. I tried a few times
to construct a test that would fail. It does affect how matches are
displayed in the playground, though, and it probably affects autofix.

Test plan:

Unfortunately we don't have automated tests for semgrep core which check
the reported column for a match. We have autofix tests, but those rely
on the offset only, and I believe they are only set up for pattern tests
anyway, not full rule tests. We do have end to end tests for the CLI,
but they are brittle and each one comes with a decent maintenance
burden. For this, I think it makes sense to just manually test.

Rule:

```
rules:
- id: test
  patterns:
    - pattern: |
        '$STR'
    - metavariable-pattern:
        metavariable: $STR
        language: generic
        pattern: |
          ... a $B c ...
    - focus-metavariable: $B
  message: test
  languages: [js]
  severity: WARNING
```

Test file:

```
const asdf = 'a b c';

const asdfasdfasdf = '\
a b c\
';
```

Test command:

`semgrep-core -l js -rules test.yaml test.js -json_nodots | jq`

Before: https://gist.github.com/nmote/77b0127105fbfe801f2049eb8a5a7ad2

After: https://gist.github.com/nmote/3091a3996dff2f5c363f625c4f4a6fc0

Note that the location of the first match is unchanged, but the location
of the second match now correctly reports the start column as 3 and end
as 4 instead of 25 to 26. The location of the metavar `$B` is similarly
updated.