fix: Reduce logspew with Pro Engine
I was seeing this a lot with `--debug` and it was drowning out other
findings and leading to excessive memory use on the CLI side (which
stores up all the output and then prints it at the end).

After this change, the CLI's memory usage with `--debug --pro` is
greatly reduced.

Test plan: Manually tested with `--debug --pro` and observed greatly
reduced memory usage.