fix: Improve parse error messages related to metavariable-pattern
Currently, when there is a nested language in a `metavariable-pattern`
and we encounter a parse error when parsing the metavariable's contents,
we get something like this (appears when providing the `--verbose` flag
to the CLI):

```
[WARN] Syntax error at line /tmp/tmp-3427277-52324a.mvar-pattern:1:
 `<` was unexpected
```

This is totally unactionable, and is a bit scary. There is no way to
tell what rule or file led to this result. To improve this situation, I
propose two changes in this pull request:

First, when we encounter errors when running a nested pattern in a
different language for metavariable-pattern, we annotate the resulting
error message with information about the nested target language and the
rule ID.

Second, when we get parse errors, we now update their locations so that
they point to the original source text rather than to the temp file
created for the metavariable pattern.

I'm not completely convinced that we should be even issuing these errors
at all. I think it is expected that the contents of a metavariable might
not always parse cleanly as another language. But, it's probably better
than silently declining to provide results because of a parse error that
users don't have any way of finding out about. I'd like to see if this
messaging change clears up confusion first and iterate if not.

Test plan:

Rule:

```
rules:
- id: find.really.scary.problems
  patterns:
    - pattern: "$X"
    - metavariable-pattern:
        metavariable: $X
        language: js
        pattern: ...
  message: placeholder message
  languages: [python]
  severity: WARNING
```

test.py:
```
y = 'asdf'

x = '<'
```

Run `semgrep scan -c test.yaml . --verbose` and observe the following lines in the output:

```
[WARN] Syntax error at line test.py:3:
 When parsing a snippet as JavaScript for metavariable-pattern in rule 'find.really.scary.problems', `<` was unexpected
```