const-prop: Let Pro engine resolve functions
In Pro, a function is assigned a unique key used for lookups, and how
this key is obtained should not matter to Semgrep OSS. So we pass the
entire function expression to the Pro hook and let the Pro Engine do
its the magic. This is how taint analysis hook works and const-prop hook
should work the same way.

Also do some code/style clean up.

Follows returntocorp/semgrep-proprietary#545

test plan:
make test