feat: Use Semgrep to enforce correct changelog extensions
Semgrep doesn't really have support for enforcing standards on path
names. The best we can do right now is the workaround described in #2438
to include/exclude paths, and then always match any contents in files
that the rule applies to.

Unfortunately, the workaround generates a separate finding for each
line, but that's really not a big deal.

Test plan:

Add files `asdf.js` and `asdf.fix` to the `changelog.d` directory. Run
`semgrep -c semgrep.yml` and observe findings both when the files are
empty and when they are populated. Otherwise, observe no findings.