Break dependency cycle between Logging and Common
There is a dependency cycle in `libs/commons`: `Common.ml` depends on
`Logging`, and `logging.mli` depends on `Common`.

Dune used to be OK with that, but the new dependency analysis algorithm
in 3.7 makes tighter checks and (correctly) detects it. It's kinda OK in
that particular case because `Logging` only uses a type alias from
`Common`, but dune can not reliably detect this.

So this PR expands the type alias in a way that splits the cycle and
makes it compatible with dune 3.7.

Signed-off-by: Etienne Millon <me@emillon.org>