refactor: Remove weird visitor code
In #7186 I kept this in so as to preserve the existing behavior as much
as possible. I wanted to do this cleanup separately so that we would
minimize the amount of code we might have to revert if this caused a
regression that tests didn't catch.

The deleted comments explain what this code was for.

Test plan: Automated tests