chore: Tweak AST_stat counting of nodes
This currently counts a bunch of nodes that are created only to
facilitate matching, but aren't present in the original AST. Of course,
this is just a loose approximation anyway since it's not clear how one
should go about counting the total number of meaningful nodes in a tree
like our generic AST, but I think it's better to be more faithful to the
AST than less.

I intend to migrate this off of the matching visitor and use a vanilla
visitor instead. I'm doing this as an intermediate step so that I can
prove that the changes in numbers that we see are due to these extra
nodes.

Test plan:

From `stats/parsing-stats`, run `./run-lang julia` (after running
`make core && make install` at the repo root).

The untranslated node count is unchanged (149476) but the total node
count drops from 9432006 to 9005661