feat(perf): Only parse rules that will actually be run
Pattern parsing--specifically, Ruby pattern parsing--can be slow.
We currently have over 1000 rules in p/default, many of which are
not run depending on the languages in the repo. We can skip the
pattern parsing stage for those rules.

In conjunction with https://github.com/returntocorp/semgrep/pull/7188,
reduces rule parsing when run on the semgrep repo  to around a second,
down from 13 seconds with just 7188. It is very effective because Ruby
pattern parsing takes the vast majority of the time. Any repo without
Ruby should see a similar result.

Curently not implemented for osemgrep.

Test plan: make test