AST_to_IL: Fix translation of side-effects in Bool exprs
We were interpreting `A && x = 42` as `x = 42; A && x`, but the side-
effects of `x = 42` only take effect if `A` holds.

We fix this by guarding these side-effects, e.g. interpreting
`A && x = 42` as `if (A) x = 42; A && x`.

Unfortunately, this will have some side-effects on sym-prop, because we
kill sym-prop values at JOIN points, so any sym-prop values are lost
after an if-statement. Sym-prop is experimental so that's OK.

Closes #7199

test plan:
make test # tests added