fix(metrics): Send metrics for app scans
We currently do not send metrics for most scans triggered by semgrep ci
due to an error in the logic. This makes it harder for us to debug issues
for our customers, since there may be useful information in the metrics
that we don't have access to. This PR fixes the bug

Test plan: I tested this by running `semgrep ci` and printing out
whether metrics were enabled. e2e tests TODO