CI: remove use of a dune cache (#7327)
Not sure what was the purpose of this caching code. I think
Bence added the code a long time ago. Maybe when chris
removed some of the semgrep-core caching code bence wrote
he forgot also to remove this bit.

Maybe this is useful when ocaml-layer lags behind,
but this step currently takes 1m, so seems simpler
to just rely on ocaml-layer for caching the needed packages
and just faster too.

test plan:
wait for CI to kick in, look at the time for the Test semgrep-core
job and look if less than 10min51s