refactor: Migrate Match_patterns.ml to new visitor
This also changes the `Matching_visitor` to inherit from
`iter_no_id_info`. Previously, the default to not recurse there was
encoded in the `kid_info` function provided in the `default_visitor`.

This subtly changes the behavior of `mk_visitor`, but no remaining
callsites depend on that behavior and I will delete it soon.

Test plan: Automated tests