Semgrep.js: start to generate JS files for a few languages
The difference between small vs big grammar and
using pfff or not does not seem to impact a lot the size
of the JS files. They are almost all at 2.5MB.
Maybe I did something wrong.
Still, it's a first step towards splitting the JS file
of the engine in multiple "bundles".

test plan:
```
$ source libs/ocaml-tree-sitter-core/tree-sitter-config.sh
$ make core
$ dune build --profile=release
$ ls -l _build/default/js/languages/*/
```