osemgrep: use http lwt client (#7400)
* osemgrep: rename core/Metrics to core/Metrics_

The opam package "metrics" defines the module Metrics. Previously, this couldn't
be used in osemgrep. Additionally, no package depending on metrics could be used.

* osemgrep: use http-lwt-client instead of cohttp-lwt-unix

Previously, neither IPv6 nor HTTP/2 were supported, and CoHTTP (with lots of
functors inside) was used for HTTP requests. Now, the functor-free opam package
http-lwt-client is used, relying on happy-eyeballs to use IPv4 or IPv6 and
httpaf for HTTP/1 and h2 for HTTP/2 support.

* osemgrep/networking: indentation