fix(autofix/python): Fix range of expressions wrapped in parens
Previously, the `$X` metavariable binding in the new fix_binop.py
example was given the range associated with `1 + (5 + 3`. This is
because the parentheses get dropped by the parser, so the location of
the second plus operation is associated with only `5 + 3`. In isolation
that's fine, but in cases like this where it is part of a larger
expression, this casn mean that one parenthesis is included in the range
while the other is not.

This issue affects both text-based autofix and AST-based autofix. Both
need to know which snippet to replace with the fix, and both need to
know the correct location of the original text for bound metavariables.

My first attempt to address this was to create `ParenExpr` nodes in the
generic AST, which are already used on a limited basis for Go. Even
after adding cases to unwrap them in `m_expr`, there was still a fairly
large difference in matching behavior. I assume that there are a lot of
places outside of `m_expr` that match on expression nodes, and we'd have
to add unwrapping logic everywhere that happens. This seemed error prone
and hard to maintain. This is another case that illustrates the tension
between an AST, most useful for matching and analysis, and a CST, most
useful for manipulating and printing code.

Instead, I decided to write the range, including the parentheses, to
`e_range` when the expression node was created. This is a bit of a
departure from the way e_range has historically been used. Until now, as
far as I know, it has only been a cache. Now, it serves as the canonical
source of range information in this case. I think that this is a good
direction to move in -- many of our location issues stem from some
tokens not making it into the resulting AST, so I think that explicitly
setting the range based on locations from the parser, when the node is
constructed, will help. That way, we would be free to drop semantically
meaningless tokens from the AST without worrying that they are needed to
compute locations.

Fixes #2902

Test plan:

- Automated tests
- Manual test with the example from #2902
  (https://semgrep.dev/playground/s/20WB)
  - Also manual test after putting the tree sitter parser first in
    Parse_target2.ml