tainting: Fix bug in field sensitivity (#7463)
* tainting: Fix bug in field sensitivity

Closes PA-2486

We have to distinguish between "clean in the environment" versus
"matches a sanitizer". If an l-value such as `x.a.b` is tainted, but
that same occurrence of `x.a` matches a sanitizer, we will not taint
`x.a.b` (the sanitizer "blocks" the taint). However, if `x.a` is just
clean in the environment from a previous sanitization, then `x.a.b`
must get tainted.

test plan:
make test # test added

* Emma's review