cleanup: get rid of cli-bridge/
This complicates the build and code and is not used anymore
(this used to be an experiment for snowflake UDF, but
even with the bridge the snowflake UDF can't execute
the code in semgrep-core (even when linked with the python
executable), so the extension is not useful anymore.

Moreover with osemgrep there is even less a need to
have a bridge with semgrep-core.

test plan:
make test