refactor: Make custom map visitor publicly accessible (#7641)
This moves the nonstandard parts of of the mapper into AST_generic. Now,
the mapper in `Map_AST`'s only responsibility is to override methods and
call the appropriate client-supplied function. This makes it possible
for callsites to start using the custom visitor directly, at which point
we can delete `Map_AST.mk_visitor`.

Test plan: Automated tests