Remove .devcontainer/ and adjust semgrep-dev dockerfile
I have no idea what this devcontainer thing is.
It was added in https://github.com/returntocorp/semgrep/pull/5131
but there is no documentation on how to use it, it has not been
updated in 1 year, so there is no way to know if it still works.

It also added pre-commit in the semgrep-dev container image which
caused some annoying issues like
https://github.com/returntocorp/semgrep/pull/7610
and https://github.com/returntocorp/semgrep/pull/7642
so simpler to just remove all of that and revert

test plan:
Wait for CI to see if it breaks anything.
Wait for outside people to complain, so we can actually
know whether this thing is still used