refactor: Move Matching_visitor to src/matching
This used to be the default behavior of the visitor that was used
everywhere, since there wasn't a reasonable way to have different
visitors with different behaviors. It's customized for matching, and
it's now only used for matching. It makes sense to colocate it with
matching code.

Test plan: Automated tests