fix(js): Avoid dropping parentheses when autofixing
This is equivalent to #7437, but for JS. Now we include parentheses in
the range for expressions, so we don't end up in weird situations where
the location of an expression spans one parenthesis but not its match.

Fixes #6233

Probably fixes a couple other issues too (will check after).

Test plan:

* Automated tests
* Manual test of the case given in #6233
* Put the Menhir parser first in `Parse_target2.ml` then manual test of
  https://semgrep.dev/s/EAYw.