refactor: Remove ParenExpr
I believe that this was only added to make sure the ranges were correct
for expressions surrounded by parentheses, especially for autofix.
However, I ran into correctness issues when I tried to use it for Python
in #7437. I assume that this is because there are places that should
match and recurse on `ParenExpr` that were not doing so. It would be
difficult to make sure that we are handling it correctly throughout
Semgrep.

Instead, I set `e_range` to include the locations of the parentheses.
This worked well, and I followed up by doing the same for JS in #7686. I
think it's time to take this approach for other languages, and simplify
the generic AST and analyses that run on it by removing this node.

Test plan: Automated tests