feat(scala): parse if expr `then` better (#7696)
## What:
This PR makes it so that we parse the `then` expression in a Scala
if-then-else better.

## Why:
Previously, we were unable to parse things such as 
```
if (true)
  x.foo();
  val x = 2
else
  4
```
because our `blockStatSeq` parser is limited to things which do not
start with expressions. This was so we had a distinction between
blockStatSeq's and just regular `expr`s, which would otherwise conflate,
and cause us to incorrectly parse things as a `blockExpr` when it's too
general.

In addition, there are difficulties with parsing it as such, because we
have things such as
```
if (true)
  3 else 4
```
where the expression is indented, which would ordinarily start an
indentation region, but there is no corresponding
DEDENT at the end of the expression. This messes with things like
`inBracesOrIndented` and friends, which expect to parse a single
nonterminal, and then parse the DEDENT. Here, however, there's not just
a single nonterminal before the DEDENT, because we have the `else 4` in
addition to the `3`.

## How:
We added the `indentexExprOrBlackStatSeqUntil` function, which means to
parse something which is an indented expression, or a block stat seq. In
addition, there is the possibility of a terminating token, which is
cnotextual. In the case of a `then` expression, this is `else`. In the
case of the terminating token, we don't require a DEDENT, so we can't
use a normal parsing combinator.

## Test plan:
Included tests, and parse rate `0.9785559783011464` ->
`0.9800450983204928`


PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)